requirements.txt
streamlit==1.28.1
azure-identity==1.14.0
azure-mgmt-datafactory==9.6.0
openai==1.3.0
python-dotenv==1.0.0
lxml==4.9.3
